requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.1.31
click==8.1.8
colorama==0.4.6
fastapi==0.115.12
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
mslex==1.3.0
packaging==24.2
pluggy==1.5.0
psutil==6.1.1
psycopg2==2.9.10
psycopg2-binary==2.9.10
pydantic==2.11.1
pydantic-core==2.33.0
pytest==8.3.5
python-dotenv==1.1.0
sniffio==1.3.1
sqlalchemy==2.0.40
starlette==0.46.1
taskipy==1.14.1
tomli==2.2.1
typing-extensions==4.13.0
typing-inspection==0.4.0
uvicorn==0.34.0
